requirements.txt
flask==2.0.1
werkzeug==2.0.2
flask-cors==3.0.10
python-dotenv==1.0.0
gunicorn==20.1.0
requests==2.31.0
ctransformers==0.2.27